Import of MLEHMANN/Coro-4.804 from CPAN.
gitpan-cpan-distribution: Coro
gitpan-cpan-version:      4.804
gitpan-cpan-path:         MLEHMANN/Coro-4.804.tar.gz
gitpan-cpan-author:       MLEHMANN
gitpan-cpan-maturity:     released

+++ Coro/Debug.pm
@@ -106,16 +106,15 @@ use Carp ();
 use Time::HiRes ();
 use Scalar::Util ();
 
-use IO::Socket::UNIX ();
-
 use AnyEvent ();
+use AnyEvent::Socket ();
 
 use Coro ();
 use Coro::Handle ();
 use Coro::State ();
 use Coro::AnyEvent ();
 
-our $VERSION = 4.803;
+our $VERSION = 4.804;
 
 our %log;
 our $SESLOGLEVEL = exists $ENV{PERL_CORO_DEFAULT_LOGLEVEL} ? $ENV{PERL_CORO_DEFAULT_LOGLEVEL} : -1;
@@ -437,12 +436,15 @@ runs C<session> on any connection. Normal unix permission checks and umask
 applies, so you can protect your socket by puttint it into a protected
 directory.
 
-The C<socat> utility is an excellent way to connect to this socket,
-offering readline and history support:
+The C<socat> utility is an excellent way to connect to this socket:
 
-   socat readline:history=/tmp/hist.corodebug unix:/path/to/socket
+   socat readline /path/to/socket
 
-The server accepts connections until it is destroyed, so you should keep
+Socat also offers history support:
+
+   socat readline:history=/tmp/hist.corodebug /path/to/socket
+
+The server accepts connections until it is destroyed, so you must keep
 the return value around as long as you want the server to stay available.
 
 =cut
@@ -451,31 +453,41 @@ sub new_unix_server {
    my ($class, $path) = @_;
 
    unlink $path;
-   my $fh = new IO::Socket::UNIX Listen => 1, Local => $path
-      or Carp::croak "Coro::Debug::Server($path): $!";
+   AnyEvent::Socket::tcp_server "unix/", $path, sub {
+      my ($fh) = @_;
+      Coro::async_pool {
+         $Coro::current->desc ("[Coro::Debug session]");
+         session $fh;
+      };
+   } or Carp::croak "Coro::Debug::new_unix_server($path): $!";
+}
 
-   my $self = bless {
-      fh   => $fh,
-      path => $path,
-   }, $class;
+=item $server = new_tcp_server Coro::Debug $port
 
-   $self->{cw} = AnyEvent->io (fh => $fh, poll => 'r', cb => sub {
-      if (my $fh = $fh->accept) {
-         Coro::async_pool {
-            $Coro::current->desc ("[Coro::Debug session]");
-            session $fh;
-         };
-      }
-   });
+Similar to C<new_unix_server>, but binds on a TCP port. I<Note that this is
+usually results in a gaping security hole>.
 
-   $self
+Currently, only a TCPv4 socket is created, in the future, a TCPv6 socket
+might also be created.
+
+=cut
+
+sub new_tcp_server {
+   my ($class, $port) = @_;
+
+   AnyEvent::Socket::tcp_server undef, $port, sub {
+      my ($fh) = @_;
+      Coro::async_pool {
+         $Coro::current->desc ("[Coro::Debug session]");
+         session $fh;
+      };
+   } or Carp::croak "Coro::Debug::new_tcp_server($port): $!";
 }
 
 sub DESTROY {
    my ($self) = @_;
 
-   unlink $self->{path};
-   close $self->{fh};
+   unlink $self->{path} if exists $self->{path};
    %$self = ();
 }
